Moved AgeGenerator into domain folder
