+++ src/sanguinehaze/charactercreator/adapter/views/MainFrame.java
@@ -1,7 +1,7 @@
 package sanguinehaze.charactercreator.adapter.views;
 
 import sanguinehaze.charactercreator.AdditionalFeatures;
-import sanguinehaze.charactercreator.AgeGenerator;
+import sanguinehaze.charactercreator.domain.AgeGenerator;
 import sanguinehaze.charactercreator.adapter.views.events.FormEvent;
 import sanguinehaze.charactercreator.GenerateSourceData;
 import sanguinehaze.charactercreator.Profession;
